+++ test/test_constraint.py
@@ -69,6 +69,12 @@ def test_empty_constraints_repr() -> None:
     Model().constraints.__repr__()
 
 
+def test_cannot_create_constraint_without_variable() -> None:
+    model = linopy.Model()
+    with pytest.raises(ValueError):
+        _ = linopy.LinearExpression(12, model) == linopy.LinearExpression(13, model)
+
+
 def test_constraints_getter(m: Model, c: linopy.constraints.Constraint) -> None:
     assert c.shape == (10,)
     assert isinstance(m.constraints[["c"]], Constraints)